organized file structured following FLAT
