+++ src/app/pokemon-details/pokemon-details.component.css
@@ -11,7 +11,7 @@
 	div#pokedex
 	{
 		width: 750px;
-		margin: 50px auto 0 auto;
+		margin: 0 auto;
 	}
 
 	div#left
@@ -224,7 +224,7 @@
 		height: 451px;
 		float: right;
 		
-		background: linear-gradient(left, #8b0000 0%, #c00d0d 50%, #8b0000 100%);	
+		background: linear-gradient(to right, #8b0000 0%, #c00d0d 50%, #8b0000 100%);	
 		background: -webkit-linear-gradient(left, #8b0000 0%, #c00d0d 50%, #8b0000 100%);
 		background: -moz-linear-gradient(left, #8b0000 0%, #c00d0d 50%, #8b0000 100%);
 		background: -o-linear-gradient(left, #8b0000 0%, #c00d0d 50%, #8b0000 100%);
@@ -257,7 +257,7 @@
 		height: 245px;
 		width: 260px;
 		padding: 0 20px;
-		background-color: #b0b0b0;
+		background-color: lightgray;
 		float:left;
 
 		position: absolute;
@@ -349,10 +349,10 @@
 		-moz-box-shadow: -2px 1px #5e0000;
 		-o-box-shadow: -2px 1px #5e0000;
 
-		background: linear-gradient(top, #c00d0d 0%, #8b0000 80%);	
-		background: -webkit-linear-gradient(top, #c00d0d 0%, #8b0000 80%);
-		background: -moz-linear-gradient(top, #c00d0d 0%, #8b0000 80%);
-		background: -o-linear-gradient(top, #c00d0d 0%, #8b0000 80%);
+		background: linear-gradient(to bottom, #c00d0d 0%, #8b0000 80%);	
+		background: -webkit-linear-gradient(to bottom, #c00d0d 0%, #8b0000 80%);
+		background: -moz-linear-gradient(to bottom, #c00d0d 0%, #8b0000 80%);
+		background: -o-linear-gradient(to bottom, #c00d0d 0%, #8b0000 80%);
 	}
 
 	div#speakers
@@ -390,10 +390,10 @@
 		-moz-border-radius: 100px;
 		-o-border-radius: 100px;
 
-		background: linear-gradient(top, #307bfb 0%, #0530e5 80%);	
-		background: -webkit-linear-gradient(top, #307bfb 0%, #0530e5 80%);
-		background: -moz-linear-gradient(top, #307bfb 0%, #0530e5 80%);
-		background: -o-linear-gradient(top, #307bfb 0%, #0530e5 80%);
+		background: linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);	
+		background: -webkit-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
+		background: -moz-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
+		background: -o-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
 
 		box-shadow: -3px 2px #001c91;
 		-webkit-box-shadow: -3px 2px #001c91;
@@ -415,10 +415,10 @@
 		-moz-border-radius: 100px;
 		-o-border-radius: 100px;
 
-		background: linear-gradient(top, #50fb05 0%, #057b05 100%);	
-		background: -webkit-linear-gradient(top, #50fb05 0%, #057b05 100%);
-		background: -moz-linear-gradient(top, #50fb05 0%, #057b05 100%);
-		background: -o-linear-gradient(top, #50fb05 0%, #057b05 100%);
+		background: linear-gradient(to bottom, #50fb05 0%, #057b05 100%);	
+		background: -webkit-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
+		background: -moz-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
+		background: -o-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
 
 		box-shadow: -1px 2px #004200;
 		-webkit-box-shadow: -1px 2px #004200;
@@ -446,10 +446,10 @@
 		-moz-border-radius: 100px;
 		-o-border-radius: 100px;
 
-		background: linear-gradient(top, #fb6505 0%, #bb0505 100%);	
-		background: -webkit-linear-gradient(top, #fb6505 0%, #bb0505 100%);
-		background: -moz-linear-gradient(top, #fb6505 0%, #bb0505 100%);
-		background: -o-linear-gradient(top, #fb6505 0%, #bb0505 100%);
+		background: linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);	
+		background: -webkit-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
+		background: -moz-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
+		background: -o-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
 
 		box-shadow: -1px 2px #7b0000;
 		-webkit-box-shadow: -1px 2px #7b0000;
@@ -799,10 +799,10 @@
 		-moz-border-radius: 10px;
 		-o-border-radius: 10px;
 
-		background: linear-gradient(top, #307bfb 0%, #0530e5 80%);	
-		background: -webkit-linear-gradient(top, #307bfb 0%, #0530e5 80%);
-		background: -moz-linear-gradient(top, #307bfb 0%, #0530e5 80%);
-		background: -o-linear-gradient(top, #307bfb 0%, #0530e5 80%);
+		background: linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);	
+		background: -webkit-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
+		background: -moz-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
+		background: -o-linear-gradient(to bottom, #307bfb 0%, #0530e5 80%);
 
 		box-shadow: -1px 2px #001c91;
 		-webkit-box-shadow: -1px 2px #001c91;
@@ -825,10 +825,10 @@
 		-moz-border-radius: 100px;
 		-o-border-radius: 100px;
 
-		background: linear-gradient(top, #50fb05 0%, #057b05 100%);	
-		background: -webkit-linear-gradient(top, #50fb05 0%, #057b05 100%);
-		background: -moz-linear-gradient(top, #50fb05 0%, #057b05 100%);
-		background: -o-linear-gradient(top, #50fb05 0%, #057b05 100%);
+		background: linear-gradient(to bottom, #50fb05 0%, #057b05 100%);	
+		background: -webkit-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
+		background: -moz-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
+		background: -o-linear-gradient(to bottom, #50fb05 0%, #057b05 100%);
 
 		box-shadow: -1px 2px #004200;
 		-webkit-box-shadow: -1px 2px #004200;
@@ -851,10 +851,10 @@
 		-moz-border-radius: 100px;
 		-o-border-radius: 100px;
 
-		background: linear-gradient(top, #fb6505 0%, #bb0505 100%);	
-		background: -webkit-linear-gradient(top, #fb6505 0%, #bb0505 100%);
-		background: -moz-linear-gradient(top, #fb6505 0%, #bb0505 100%);
-		background: -o-linear-gradient(top, #fb6505 0%, #bb0505 100%);
+		background: linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);	
+		background: -webkit-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
+		background: -moz-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
+		background: -o-linear-gradient(to bottom, #fb6505 0%, #bb0505 100%);
 
 		box-shadow: -1px 2px #7b0000;
 		-webkit-box-shadow: -1px 2px #7b0000;
