+++ src/app/pokemon-list/pokemon-list.component.css
@@ -30,7 +30,7 @@ a {
 }
 
 a:hover {
-    font-size: 1.9vw;
+    font-size: 2.3vw;
     color: #ffcb05;
 }
